Rewrote some of the architecture to support parsed bwd predictions and refactored some ancient code

+++ src/Transforms/Algorithms/BWD/Entities/BWDictionary.cs
@@ -3,13 +3,14 @@ using System.Collections.Generic;
 
 namespace BWDPerf.Transforms.Algorithms.BWD.Entities
 {
-    public class BWDictionary : Dictionary<int, ReadOnlyMemory<byte>>
+    public class BWDictionary : Dictionary<ushort, ReadOnlyMemory<byte>>
     {
         public ReadOnlyMemory<byte> Serialize()
         {
             var buffer = new List<byte>(capacity: this.Count * 3);
+            // TODO: This can be 2 bytes for ushort
             buffer.AddRange(BitConverter.GetBytes(this.Count));
-            for (int i = 0; i < this.Count; i++)
+            for (ushort i = 0; i < this.Count; i++)
             {
                 buffer.Add((byte) this[i].Length);
